pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlirAgent"]

[project]
name = "mlirAgent"
version = "0.1.0"
description = "AI-guided MLIR/LLVM compiler optimization framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0,<2",
    "python-dotenv>=1.0.0",
    "tqdm>=4.60",
    "pyyaml>=6.0",
    "neo4j>=5.0",
    "pydriller>=2.0",
    "protobuf>=4.0",
]

[project.optional-dependencies]
evolve = [
    "optuna>=3.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
mlirAgent = "mlirAgent.cli:main"
mlirAgent-mcp = "mlirAgent.mcp_server:main"
